Move g1_molmo_port's G1Robot into molmo_spaces/robots/g1_ported.py
Copies the port's G1Robot (the reference-equivalent implementation) into
molmo_spaces proper, where it will be reshaped toward robots/abstract.py's
Robot interface. components/robot_g1ms.py becomes a pure re-export shim so
every consumer shares one class object and one set of constants.

Verified byte-identical to the recorded ported-rollout baseline (58/58 trace
lines; steps=2338 sim_time=12.04s success=True on seed 0), which matches the
gold run on every discrete invariant.
